src/cassandra_demo/requirements.txt
cassandra-driver
prettytable
requests
pyopenssl
python-dotenv
cqlsh
prettytable
azure-identity
azure-storage-blob